requirements.txt
opencv-python
Pillow
torchsummary
torchvision
matplotlib
pydicom
scikit-image